Create Dice Roller in Java
This is a dice rolling program that takes user input for sides, and number of dice to roll. It then passes these dice into randomized array, sorts it through a very basic bubble sort. After that it will work through several other methods to sum, max,min,average,median. It then prompts a user for a number to search for, using both a linear and binary search and then displaying the index at which it discovers the first occurrence. The code is structured to repeat until 0 or a negative number is entered. Java has a lot of these methods built in, but these helped me understand what they're doing.